+++ src/SoftDeleteable/Event/PostSoftDeleteEventArgs.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Doctrine Behavioral Extensions package.
+ * (c) Gediminas Morkevicius <gediminas.morkevicius@gmail.com> http://www.gediminasm.org
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Gedmo\SoftDeleteable\Event;
+
+use Doctrine\Persistence\Event\LifecycleEventArgs;
+use Doctrine\Persistence\ObjectManager;
+
+/**
+ * Post soft-delete lifecycle event
+ *
+ * @extends LifecycleEventArgs<ObjectManager>
+ */
+final class PostSoftDeleteEventArgs extends LifecycleEventArgs
+{
+}